move on-prem config installer into onprem-ke installer
